Port negative dentries limit feature from 3.10
This ports the Redhat feature forward from the 3.10 kernel version.

This feature allows one to specifiy a loose maximum of total memory
which is allowed to be used for negative dentries.  This is done
via setting a sysctl variable which is used to calculate a
negative dentry limit for the system.  Every 15 seconds a kworker
task will prune back the negative dentries that exceed the limit,
plus an extra 1% for hysteresis purposes.

Intent is that the feature code is kept as close to the 3.10 version
as possible.

Main differences from the 3.10 version of the code:
- count of dentries associated with a superblock is kept in a
different location, requiring a procedure call to obtain
- superblocks are now kept by node id and memcg, requiring
more calls into iterate_super

Verification
- ensure the sysctl variable is set to 20:
sysctl fs.negative-dentry-limit
- run a test program that continuously tries to access a lot of
non-existent files, causing a lot of negative dentries to build up
- monitor the number of negative dentries building up in the system:
cat /proc/sys/fs/dentry-state
the fifth number is the negative dentries
- get the calculated negative dentry limit:
dmesg | grep Negative
- watch the negative dentries you are monitoring periodically drop
to below the limit before they start building back up again.  The
drop should happen about 4 times per minute

Closes-Bug: 2017703

Signed-off-by: Jim Somerville <jim.somerville@windriver.com>
Change-Id: I3f55249aab45471802d123ed2253b6f36cc4af50