Drop the components of livepatch feature
As there is no plan to support livepatch feature in StarlingX community,
now we drop the packages and components of this feature.
Since this component depends on the userspace tools kpatch, so this
patch should be merged first(Another commit in integ repo will depend on
this).

TestPlan:
PASS: build-pkgs -a
PASS: build-image
PASS: Jenkins installation.

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
Change-Id: Iaf96daddab40f87d6155333eae7d1780a3696764