Add pkgs without abiname for image/headers
Add 2 packages linux-image-stx-amd64/linux-headers-stx-amd64
which don't have abiname in their names. They depend on packages with
abiname in names. Then we can use these 2 packages in anywhere
that involves image/headers packages (e.g. Build-Depends/yaml config
and so on). When the abiname is changed later in any kernel upgrading
we don't need change above places involved any more.

We don't use the linux-image-amd64/linux-headers-amd64 as Debian does
because they are built by linux-signed-amd64, and coupled with signed
kernel. We don't follow Debian's signing process so we create 2 new
packages which are coupled with unsigned image.

BTW, rename package "linux-keys-@version@" to "linux-keys" because the
"@version@" isn't necessary for this package. Then the version numbers
can be completely removed from the Build-Depends of OOT kernel
modules' codes.

All of above are done on rt kernel too.

This is a preparation for kernel upgrading with major version.

Test plan:
 PASS: 2 new pkgs linux-image-stx-amd64/linux-headers-stx-amd64 can be
       built successfully for linux.
 PASS: 2 new pkgs linux-rt-image-stx-amd64/linux-rt-headers-stx-amd64
       can be built successfully for linux-rt.

Story: 2010643
Task: 48815

Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: I63f968d3b24728b2b5b08e889c26c3c4f6a0e1df