kernel: Enable the in-tree drivers and disable OOT drivers
We had updated the Linux kernel to the latest 6.6.x stable kernel.  As
part of this we want to transition from out-of-tree (OOT) kernel drivers
to in-tree drivers to reduce the development effort going forward and to
benefit from community testing and the latest bug fixes.

Enable the in-tree drivers ice, iavf, i40e, bnxt_re, bnxt_en, mlx5_core,
mlx5_ib, rpcrdma, irdma and infiniband related modules, also enable
ionic and sfc. This commit also disables all OOT drivers before we
support them again because there are build issues.

In addition, if we need support rdma function for the in-tree bnxt_re
we will use
https://github.com/linux-rdma/rdma-core/tree/master/providers/bnxt_re
Currently libbnxt_re comes from Broadcom bcm_220.0.83.0.tar.gz
in the starlingx. This part will be done in the future depend on which
driver will be used between the OOT and the in-tree drivers.

Note: The intel-ice package is preserving, but we do not enable the OOT
ice driver, only kept ddp file so that the in-tree driver ice can use
the ddp file. We will add the ddp file for in-tree driver once Intel
finished their testing.

Verification:
* ice, i40e, iavf:
  - installs from iso succeed on servers with ice(ntel E810-2C-QDA2
    Chapman beach) and i40e hw(Intel Ethernet Controller X710) for
    rt and std.
  - interfaces are up and pass packets for rt and std.
  - create vfs, ensure that they are picked up by the new iavf
    driver and that the interface can come up and pass packets
    on rt and std system.
  - Check dmesg to see DDP package is loaded successfully and
    the version is 1.3.30.0 for rt and std.
* bnxt_en/bnxt_re:
  - Installation of the ISO image is successful with standard and
    low-latency profiles.
  - Physical function interfaces are up and pass packets for rt and std.
  - Create vfs, ensure that the interface can come up and pass packets.
  - RDMA/Infiniband over Ethernet functionalities of the Broadcom
    adapters were successfully tested using the Linux RDMA community's
    perftest package.
* mlx5_core related drivers:
  - Build package success for rdma-core, mstflint.
  - Install onto a StarlingX system with All-in-One lab with network
    adapters Mellanox's OFED. The network adapters of controllers
    are Mellanox Technologies MT27800 Family [ConnectX-5].
  - Install onto a StarlingX system in labs with network
    adapters Mellanox's OFED. The network adapters of controllers
    are [ConnectX-6 LX] failed. It looks not the driver issue.
  - The physical function interfaces are up and pass packets for rt
    and std.
  - create vfs, ensure that the interface can come up and pass packets.
  - RDMA/Infiniband over Ethernet functionalities of the Mellanox
    adapters were successfully tested using the Linux RDMA community's
    perftest package.

Story: 2011056
Task: 49630

Change-Id: Icb9e6c111532edf03c979144f8ce32a81006834f
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
(cherry picked from commit e8e3c290fc64eb3a5f08ab624683a1c7afb43892)
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>