Adapt kernel 6.6 from bullseye to bookworm platform
We have upgraded to kernel 6.6.7 on bullseye (Debian 11) platform
(on branch f/kernel-6.6) now. Because the work about upgrading to
bookworm (Debian 12) begins, the kernel 6.6 on bullseye need to be
adapted to work on bookworm, or else build errors occur such as
unmatched gcc version.
Update the version of Debian upstream "debian" folder from
6.1.27-1~bpo11+1 (version for porting back 6.1.27 from bookworm
to bullseye) to 6.1.27-1 (version for 6.1.27 on bookworm).
Minor rebase changes of deb_patches are needed for context changes
because 6.1.27-1~bpo11+1 and 6.1.27-1 are close.

Test plan:
 Because the bookworm distro for StarlingX debian hasn't been setup
 completely, the test of kernel only includes building now.
 PASS: build-pkgs -c -p linux,linux-rt

Story: 2011048
Task: 49752

Change-Id: I3b4c44fbf6cf59aafd09e0ffa24d39eb0c41fe34
Signed-off-by: Li Zhou <li.zhou@windriver.com>
`