kernel-modules: re-enable qat and octeon_ep OOT drivers
The drivers qat and octeon_ep are known to only exist as OOT,
because the in-tree drivers of 6.6.7 kernel do not support the
virtual function for both qat and octeon_ep.

We encountered a few version compatibility issues after upgrading
the kernel to 6.6.7 version. We adapted drivers' code to the kernel
6.6.7 by referring to the upstream commits.

We also upgrade the qat to the version 1.1.40-00018 from 1.0.20-00008,
because there is a risk if we adapted the qat-1.0.20-00008 driver to
the kernel 6.6.7. The code changes involve the changes of the
initialization logic and also involve the several files changes
because that the function of devm_device_remove_group() has been
removed in kernel 6.6.7.

Verification:
 * An ISO image can be built with this commit applied to a repo
   project of StarlingX's kernel 6.6.7 branch.
 * The ISO image can be installed to a Dell XR11 server with a DORA
   card and a server with Intel QuickAssist Technology Hardware Version,
   and the system is successfully Ansible-bootstrapped.
 * The octeon_ep driver can be loaded manually with modprobe, and a
   PF interface is instantiated by the kernel and can transfer packets.
 * Virtual function (VF) interfaces can be brought up and transfer
   packets.
 * qat_service status/start/shutdown works well.
 * cpa_sample_code can run successfully.

Story: 2011056
Task: 49775

Change-Id: Icf19340396d99bde97533bfecf22472e01f6be34
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
(cherry picked from commit 8bdd6fe0d0d1adee7516a38d851bd2457ac895a9)
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>