Add driver versions for ice, i40e and iavf
As we need the driver version to track the driver information, add
the necessary driver version for the Intel drivers.

We will use a suffix approach (like "6.6.7-stx.X", where we bump
the "X" for every local patch, 6.6.7 is the kernel version).
The first version will be ice-6.6.7-stx.0.ko.

If we upgrade the kernel from v6.6.7 to v6.6.12, and if all the
local patches are already included in 6.6.12 release, then the
driver version should reflect to ice-6.6.12-stx.0.ko, because
there is no local patch. If there are still patches that are not
part of the upstream kernel, and we have to port and include
those local patches in, then it will be ice-6.6.12-stx.X.ko(where
X=1). Since we rebase to 6.6.12, our baseline driver code is 6.6.12,
so any changes merged after that baseline, will restart the patch
number from 1 to reflect as such.

Verification:
- installs from iso succeed on servers.
- Check the driver version with dmesg, ethtool and modinfo
  for ice, i40e and iavf, it is correct 6.6.7-stx.0. 6.6.7-stx.0
  is the first version number for these three drivers.

Story: 2011056
Task: 50000

Change-Id: I3c7e1c0064fecc75d4bc18eed169f27e51e51220
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>