kernel:ice: Use irq_update_affinity_hint
This commit makes the ice device driver use the irq_update_affinity_hint
function instead of the irq_set_affinity_hint function. This is done
because the latter function sets the IRQ CPU affinities, whereas the
former does not, and this allows the use of the default IRQ affinity CPU
mask provided via the irqaffinity= kernel command line option.

Please note that this patch was not cherry-picked from an upstream
commit. The changes have been inspired by the i40e and iavf device
driver patches in the following patch series:
  https://lore.kernel.org/netdev/20210903152430.244937-1-nitesh@redhat.com/t/#u

We only need do the modification for the ice driver, the related changes for
i40e and iavf already were included in kernel 6.6.7.

We also change the ice version to 6.6.7-stx.1 from 6.6.7-stx.0 because
we do the code change for ice driver.

Verification:
- Install onto a StarlingX system with All-in-One lab.
- Check the smp_affinity_list of the ice interface related interrupts,
  the values are correct.

Story: 2011056
Task: 50075

Change-Id: I3ff1e7482d22e20da4565523b0fb737eab2b9899
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>