Set a build size advisory
The StarlingX package builder is moving to support tmpfs
build environments.  In preparation for this, we need to
add build size advisories to the control files of some packages
that need abnormally large ammounts of space in which to build.
If the package is too big to build in a tmpfs environment, a
disk based environment will be used instead.

Partial-Bug: 2081843
Change-Id: I4246f8303cae03891319922bc63804d5fa38c401
Signed-off-by: Scott Little <scott.little@windriver.com>