sched/fair: Make the BW replenish timer expire in hardirq context for PREEMPT_RT
We have the deadlock problem in the following situation:

Task relationship:
"ktimers/6"-->"irq/146-ice-enp"--
|                                |
|             "irq/165-ice-enp"-->"containerd"-->"unpigz"--
|                                |                        |
|                  ......        |                        |
|             "containerd-shim"--                         |
|                                                         |
-----------------------------------------------------------

"containerd" on cpu 38 try to get the write lock, but can not get it
because the read lock is hold by "unpigz" on cpu6. The total 12
tasks are waiting for the rt_mutex that hold by "containerd" on cpu38.

"unpigz" on cpu6 is running status, but it did not run for the 5
minutes, so it caused the last rcu panic. because the group that
included "unpigz" is entering throttled state before 5 minutes.
so the group 37 running tasks can not run during these 5 minutes.

Why wasn't the group unthrottled? "ktimers/6" is waiting for the
rtmutex that hold by "irq/146-ice-enp" on cpu6, "irq/146-ice-enp"
blocked by "containerd"on cpu38 , "containerd" blocked by "unpigz"
cpu6 end up throttled. it is a deadlock. "ktimers/6" can not run to
exec the cfs_bandwidth.period_timer to replenish runtime to
unthrottle the group.

There is a series of patches in
https://lore.kernel.org/all/20250409120746.635476-1-ziqianlu@bytedance.com/
We have the same problem as the link describes. When these patches get
merged into 6.12 we will need to revert this commit.

Link: https://lore.kernel.org/all/xhsmhttqvnall.mognet@vschneid.remote.csb/

TestPlan:
PASS: build-pkgs && build-image
PASS: Jenkins Installation

Closes-Bug: 2111309

Change-Id: I06b5ae55b513e02ca60313f15093a4782dfa3fb1
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Signed-off-by: Wentao Zhang <wentao.zhang@windriver.com>