ice: Remove deprecated ice_lag_move_new_vf_nodes() call
This a hotfix to a kernel crash that occurs on ice driver,
based on commit 34138ea02a608dc39f04e364f3249e12f6002bad
from mainline linux kernel branch

Moving the code to handle the LAG part of a VF reset to helper
functions deprecated the function ice_lag_move_new_vf_nodes().
The cleanup missed a call to this function in the error path of
ice_vc_cfg_qs_msg().

In the case that would end in the error path, a NULL pointer would
be encountered due to the empty list of netdevs for members of the
aggregate.

Remove the unnecessary call to ice_lag_move_new_vf_nodes().

Closes-Bug: 2126786

Test Plan:
PASS: Build ISO and fresh install
PASS: Soaking test

Change-Id: I3221b5741092022f87e9cae0ff0ff61889f71562
Signed-off-by: Alyson Deives Pereira <alyson.deivespereira@windriver.com>